Flyttat alla tester till separata klasser.
