+++ xUnitTests/Frontend/SignUpValidatorTests.cs
@@ -1,96 +1,23 @@
 ﻿using Frontend.Models;
 using SocialNetwork.Classes.Frontend;
 
-namespace xUnitTests
+namespace xUnitTests.Frontend
 {
-    public class TestFrontEnd
+    public class SignUpValidatorTests
     {
-        #region PasswordValidator
-
-        [Fact]
-        public void ShouldBeAbleToUseValidPassword()
-        {
-            var result = PasswordValidator.Validate("Password123");
-
-            Assert.True(result.Success);
-        }
-        
-        [Theory]
-        [InlineData("password123")]
-        [InlineData("Password")]
-        [InlineData("Passw")]
-        public void ShouldNotBeAbleToUseInvalidPassword(string password)
-        {
-            var result = PasswordValidator.Validate(password);
-
-            Assert.False(result.Success);
-        }
-
-        [Fact]
-        public void ShouldBeAbleToUsePasswordWithSixCharacters()
-        {
-            var result = PasswordValidator.CheckLength("Length");
-
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void ShouldNotBeAbleToUsePasswordWithLessThanSixCharacters()
-        {
-            var result = PasswordValidator.CheckLength("Short");
-
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void ShouldBeAbleToUsePasswordWithUpperCase()
-        {
-            var result = PasswordValidator.CheckForUpperCase("Password");
-
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void ShouldNotBeAbleToUsePasswordWithoutUpperCase()
-        {
-            var result = PasswordValidator.CheckForUpperCase("password");
-
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void ShouldBeAbleToUsePasswordWithNumeric()
-        {
-            var result = PasswordValidator.CheckForNumeric("Password1");
-
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void ShouldNotBeAbleToUsePasswordWithoutNumeric()
-        {
-            var result = PasswordValidator.CheckForNumeric("Password");
-
-            Assert.False(result);
-        }
-
-        #endregion
-
-        #region SignUpValidator
-
         [Fact]
         public void ShouldBeAbleToRegisterUserWithValidForm()
         {
             var registerModel = new RegisterModel
             {
                 FirstName = "First",
-                LastName ="Last",
+                LastName = "Last",
                 Username = "Username",
                 Email = "test@mail.com",
                 Password = "Password123",
                 RepeatPassword = "Password123"
             };
-            
+
             var result = SignUpValidator.Validate(registerModel);
 
             Assert.True(result.Success);
@@ -227,54 +154,5 @@ namespace xUnitTests
 
             Assert.False(result);
         }
-
-        #endregion
-
-        #region SignInValidator
-
-        [Fact]
-        public void ShouldBeAbleToLogInUsingValidUsername()
-        {
-            var result = SignInValidator.CheckUsername("Username");
-
-            Assert.True(result);
-        }
-
-        [Theory]
-        [InlineData("")]
-        [InlineData("U")]
-        [InlineData("Us")]
-        [InlineData("Use")]
-        [InlineData("User")]
-        public void ShouldNotBeAbleToLogInUsingInvalidUsername(string username)
-        {
-            var result = SignInValidator.CheckUsername(username);
-
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void ShouldBeAbleToLogInUsingValidPassword()
-        {
-            var result = SignInValidator.CheckPassword("Password123");
-
-            Assert.True(result);
-        }
-
-        [Theory]
-        [InlineData("")]
-        [InlineData("P")]
-        [InlineData("Pa")]
-        [InlineData("Pas")]
-        [InlineData("Pass")]
-        [InlineData("Passw")]
-        public void ShouldNotBeAbleToLogInUsingInvalidPassword(string password)
-        {
-            var result = SignInValidator.CheckPassword(password);
-
-            Assert.False(result);
-        }
-
-        #endregion
     }
 }
